backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
ultralytics==8.2.0
opencv-python==4.8.1.78
mysql-connector-python==8.1.0
bcrypt==4.0.1